Moving to the LDTTeam Package and publishing the utilities jar.
Added Utilities:
Utilities will contain systems and classes that make the work with the core
BO engine simpler and faster. To facilitate this a lot of consumers for the builders
have now been turned into var args meaning that multiple consumers can be passed in,
they will be treated as if they were executed in one (giving the user two options:
either use method chaining and one callback, or break the chain and use multiple).
Newly added utilities should however stay generic. Meaning that all parameters should be passed in and nothing should be assumed or locked in, unless it is a minecraft constant.